+++ builtin/providers/circonus/resource_circonus_check_http.go
@@ -38,13 +38,13 @@ var _CheckHTTPDescriptions = _AttrDescrs{
 	_CheckHTTPAuthPasswordAttr: "The HTTP Authentication user password",
 	_CheckHTTPAuthUserAttr:     "The HTTP Authentication user name",
 	_CheckHTTPBodyRegexpAttr:   `This regular expression is matched against the body of the response. If a match is not found, the check will be marked as "bad.`,
-	_CheckHTTPCAChainAttr:      "A path to a file containing all the certificate authorities that should be loaded to validate the remote certificate (for SSL checks)",
+	_CheckHTTPCAChainAttr:      "A path to a file containing all the certificate authorities that should be loaded to validate the remote certificate (for TLS checks)",
 	_CheckHTTPCodeRegexpAttr:   `The HTTP code that is expected. If the code received does not match this regular expression, the check is marked as "bad."`,
-	_CheckHTTPCiphersAttr:      "A list of ciphers to be used in the SSL protocol (for SSL checks)",
+	_CheckHTTPCiphersAttr:      "A list of ciphers to be used in the TLS protocol (for HTTPS checks)",
 	_CheckHTTPCertFileAttr:     "A path to a file containing the client certificate that will be presented to the remote server (for TLS-enabled checks)",
 	_CheckHTTPExtractAttr:      "This regular expression is matched against the body of the response globally. The first capturing match is the key and the second capturing match is the value. Each key/value extracted is registered as a metric for the check.",
 	_CheckHTTPHeadersAttr:      "Map of HTTP Headers to send along with HTTP Requests",
-	_CheckHTTPKeyFileAttr:      "A path to a file containing key to be used in conjunction with the cilent certificate (for SSL checks)",
+	_CheckHTTPKeyFileAttr:      "A path to a file containing key to be used in conjunction with the cilent certificate (for TLS checks)",
 	_CheckHTTPMethodAttr:       "The HTTP method to use",
 	_CheckHTTPPayloadAttr:      "The information transferred as the payload of an HTTP request",
 	_CheckHTTPReadLimitAttr:    "Sets an approximate limit on the data read (0 means no limit)",
